Started to implement the Gesture Detector, does not work yet
Gesture Detector in HorizontalMindmapView does not seem to receive the
onDown event, and the onFling event always has event1==null. This is a
problem. It might be caused by the MainActivity which processes the
onItemClicked event itself, and maybe thus does not forward the onDown
event to the horizontal mindmap view.