Started to extend LinearLayout in MindmapNode
the idea is that I can use MindmapNode directly as element in the
NodeColumn adapter. However, at the moment, it is not really possible to
load the mindmap_node_list_item.xml (by inflating it), because then the
node-specific details are not set, and when creating a node (without
inflating), the layout is not there.

I will have to find out how to get a MindmapNode, which inflates its
layout from the XML layout file, and still has all the stuff set that we
need.